fix Dagger NoSuchMethodError in current SNAPSHOT
Dagger's `dagger.internal.Preconditions.checkNotNullFromProvides` is pretty new.  If an old version of Dagger is in the build classpath of the target project, dependency resolution may resolve to that old version and cause a `NoSuchMethodError`.

That new method is used in the Provider factory for static `@Provides` functions in `@Module` objects.

Since I actually really hate qualifiers anyway, the chosen band-aid is to just create strong Provider types with injected constructors for each build property we need.