clean up unused repository declarations
This is mostly for the sake of (significantly) faster dependency resolution during the `dependencyUpdates` task.  Having the snapshots repository wide open -- when it isn't even used -- means that the Ben Manes plugin will resolve every single snapshot version of every dependency.  This is mostly a problem with Kotest, which can take as much as 10 minutes to resolve every single version/artifact.