split up long-running Actions jobs for caching purposes
If a job times out or just gets cancelled, `gradle-build-action` won't cache anything.  However, if an upstream task using the same action was already completed, that one *will* get cached.

So if for example builds/dokka/tests are split into different stages, and the last one gets cancelled, the results of the first two will still be cached.  This means that subsequent attempts will be faster.