revert the `project.isRootOfComposite()` check for matrix-yaml
The intent was to make sure that the plugin was only applied to ModuleCheck's root.  Unfortunately, the super-strict `isRootOfComposite()` check also makes it so that the project can never be included in a downstream consumer project.  With this check, if MC is added as an included build, every Gradle invocation fails.