use an `all-checks` capstone job in CI
The `all-checks` GitHub Actions job waits for all other jobs to finish, then asserts that they all passed or were skipped.  If anything fails or is cancelled, then it fails.

With this change, the GitHub branch protection settings can be changed so that this job is required, and all the other jobs can be removed from requirements.

As a result, the job list can be modified (added, removed, renamed) within the `ci.yml` file and the results will take effect immediately when they're PRed, without ever needing to modify the settings on GitHub.